Added checks for files
If a certain file extension doesn't exist, then we don't worry about deleting it.

+++ living/clear.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
-rm *.aux
-rm *.log
-rm *.pdf
-rm *.toc
+if [ -f *.aux ] ; then
+  rm *.aux
+fi
+
+if [ -f *.log ] ; then
+  rm *.log
+fi
+
+if [ -f *.pdf ] ; then
+  rm *.pdf
+fi
+
+if [ -f *.toc ] ; then
+  rm *.toc;
+fi
